--- src/components/infopanel/components/flereopplysninger/flereopplysninger-tekster.ts
@@ -1,22 +1,0 @@
-const FlereOpplysningerTekster: any = {
-    nb: {
-        'arbeidsevne.tittel': 'Hva skal til for å bedre arbeidsevnen?',
-        'arbeidsevne.tilrettelegging.tittel': 'Tilrettelegging/hensyn som bør tas på arbeidsplassen',
-        'arbeidsevne.tiltak.tittel': 'Tiltak i regi av NAV',
-        'utdypende.tittel': 'Utdypende opplysninger',
-        'friskmelding.tittel': 'Friskmelding/prognose',
-        'friskmelding.arbeidsfor.tittel': 'Pasienten er 100 prosent arbeidsfør etter denne perioden',
-        'friskmelding.hensyn': 'Beskriv eventuelle hensyn som må tas på arbeidsplassen',
-        'friskmelding.samme-arbeidsgiver': 'Jeg antar at pasienten på sikt kan komme tilbake til samme arbeidsgiver',
-        'friskmelding.arbeidfom': 'Anslå når du tror dette kan skje',
-        'friskmelding.annen-arbeidsgiver': 'Jeg antar at pasienten på sikt kan komme i arbeid hos annen arbeidsgiver',
-        'friskmelding.vurderingsdato': 'Når antar du å kunne gi tilbakemelding på dette?',
-        'friskmelding.ingen-arbeidsgiver': 'Jeg antar at pasienten på sikt kan komme tilbake i arbeid',
-        'friskmelding.ingen-arbeidsgiver.vurdering': 'Når antar du å kunne gi tilbakemelding på dette?',
-    },
-    en: {},
-};
-
-const tekster = FlereOpplysningerTekster.nb;
-
-export default tekster;
