+++ src/components/infopanel/components/layout/EkspanderbartPanel.tsx
