+++ src/components/infopanel/components/layout/OpplysningerSeksjon.tsx
