+++ src/components/infopanel/components/layout/ekspanderbartpanel.less
