+++ src/components/infopanel/components/panelelementer/PrognoseSeksjon.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+import tekster from '../../infopanel-tekster';
+import { Prognose } from '../../../../types/sykmeldingTypes';
+import EtikettMedTekst from '../layout/EtikettMedTekst';
+
+interface PrognoseSeksjonProps {
+    prognose?: Prognose;
+}
+
+const PrognoseSeksjon = ({ prognose }: PrognoseSeksjonProps) => {
+    if (!prognose || !prognose.hensynArbeidsplassen) {
+        return null;
+    }
+    return <EtikettMedTekst tittel={tekster['prognose.hensyn.tittel']} tekst={prognose.hensynArbeidsplassen} margin />;
+};
+
+export default PrognoseSeksjon;
