+++ src/components/infopanel/components/panelelementer/periode/SykmeldingPerioder.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import { Periode } from '../../../../types/sykmeldingTypes';
-import { sorterPerioderEldsteFoerst } from '../../../../utils/sorterSykemeldingUtils';
-import PanelRad from '../PanelRad';
+import { Periode } from '../../../../../types/sykmeldingTypes';
+import { sorterPerioderEldsteFoerst } from '../../../../../utils/sorterSykemeldingUtils';
 import PeriodeSeksjon from './PeriodeSeksjon';
+import Margin from '../../layout/Margin';
 
 interface SykmeldingPerioderProps {
     perioder: Periode[];
@@ -14,9 +14,9 @@ const SykmeldingPerioder = ({ perioder }: SykmeldingPerioderProps) => {
     return (
         <>
             {sortert.map((periode, index) => (
-                <PanelRad key={index.toString()}>
+                <Margin key={index.toString()}>
                     <PeriodeSeksjon periode={periode} underlined={sortert.length > 1} />
-                </PanelRad>
+                </Margin>
             ))}
         </>
     );
