+++ src/components/infopanel/components/panelelementer/periode/periodeseksjon.less
